requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
python-multipart
google-api-python-client
google-auth
google-auth-httplib2
requests